+++ src/store/assetInfoStore.ts
@@ -289,7 +289,7 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
 
     address: string
   ) => {
-    const { assets, balances } = get();
+    const { assets } = get();
     if (!assets || !provider) return;
 
     try {
@@ -324,14 +324,14 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
           {} as Record<string, BigNumber | undefined>
         );
 
-      set({ balances: { ...balances, ...btcBalance } });
+      set({ balances: { ...get().balances, ...btcBalance } });
     } catch {
       /*empty*/
     }
   },
 
   fetchAndSetStarknetBalance: async (address: string) => {
-    const { assets, balances } = get();
+    const { assets } = get();
     if (!assets) return;
 
     const starknetAsset = Object.values(assets).find((asset) =>
@@ -348,31 +348,32 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
       starknetAsset.tokenAddress
     );
     starknetBalance[orderPair] = new BigNumber(balance);
-    set({ balances: { ...balances, ...starknetBalance } });
+    set({ balances: { ...get().balances, ...starknetBalance } });
   },
 
   fetchAndSetSolanaBalance: async (address: PublicKey) => {
-    const { assets, balances } = get();
+    const { assets } = get();
     if (!assets) return;
 
-    const solanaAsset = Object.values(assets).find((asset) =>
+    const solanaAssets = Object.values(assets).filter((asset) =>
       isSolana(asset.chain)
     );
 
-    if (!solanaAsset) return;
+    if (!solanaAssets.length) return;
     const solanaBalance: Record<string, BigNumber | undefined> = {};
 
-    const balance = await getSolanaTokenBalance(address, solanaAsset);
-    const orderPair = getOrderPair(solanaAsset.chain, solanaAsset.tokenAddress);
+    for (const asset of solanaAssets) {
+      const balance = await getSolanaTokenBalance(address, asset);
+      const orderPair = getOrderPair(asset.chain, asset.tokenAddress);
 
-    if (isSolanaNativeToken(solanaAsset.chain, solanaAsset.tokenAddress)) {
-      const gas = 0.00380608;
-      solanaBalance[orderPair] = new BigNumber(
-        Math.max(0, Number((Number(balance) - gas).toFixed(8)))
-      );
-    } else solanaBalance[orderPair] = new BigNumber(balance);
-
-    set({ balances: { ...balances, ...solanaBalance } });
+      if (isSolanaNativeToken(asset.chain, asset.tokenAddress)) {
+        const gas = 0.00380608;
+        solanaBalance[orderPair] = new BigNumber(
+          Math.max(0, Number((Number(balance) - gas).toFixed(8)))
+        );
+      } else solanaBalance[orderPair] = new BigNumber(balance);
+    }
+    set({ balances: { ...get().balances, ...solanaBalance } });
   },
 
   clearBalances: () =>