+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.0.12]
+
+### Added
+
+- Added support for solana spl tokens.
+
+### Fixed
+
+- Fetch balance on mount.
+
 ## [1.0.11]
 
 ### Added